src/main/java/com/vamae/authorization/service/UserService.java
package com.vamae.authorization.service;

import com.vamae.authorization.model.User;
import com.vamae.authorization.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.NoSuchElementException;

@Service
@RequiredArgsConstructor
public class UserService {

    private final UserRepository userRepository;

    public int changeBalance(String username, int offset) {
        User user = findUserByUsername(username);
        user.setBalance(user.getBalance() + offset);
        userRepository.save(user);

        return user.getBalance();
    }

    public User findUserByUsername(String username) {
        return userRepository.findUserByUsername(username)
                .orElseThrow(() -> new NoSuchElementException("User not found!"));
    }
}
